requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3

psycopg[binary]

django-filter
python-dotenv

drf-spectacular

Pillow

celery>=5.4
redis>=5.0
django-redis

sentry-sdk

django-silk

pytest>=7.0
pytest-django>=4.5

sentry-sdk[django]

rollbar>=1.0.0